Sending Text Messages and beginning of adapter
- ChatMessagesAdapter added:
Not implemented yet. Going to be multipurpose in form of both displaying text messages as well as image messages.

- IMessage added:
All message types are going to inherit from this. Makes it easier to display in the future RecyclerView.

- TextMessage added:
Standard text message model. Must receive a userId and String to be able to send.

+++ app/src/main/java/com/example/awesomechatroomz/models/ChatRoom.java
@@ -3,7 +3,7 @@ package com.example.awesomechatroomz.models;
 import java.util.List;
 
 public class ChatRoom {
-    private int lastUpdated;
+    private long lastUpdated;
     private String name;
     private List<String> usersId;
     private String description;
@@ -12,14 +12,14 @@ public class ChatRoom {
         this.usersId = usersId;
     }
 
-    public int getLastUpdated() {
+    public long getLastUpdated() {
         return lastUpdated;
     }
 
-    public void setLastUpdated(int lastUpdated) {
+    public void setLastUpdated(long lastUpdated) {
         this.lastUpdated = lastUpdated;
     }
-    
+
     public String getName() {
         return name;
     }
